fix transaction and test fixtures refactor
