+++ cashews/wrapper/transaction.py
@@ -1,31 +1,19 @@
-import asyncio
-import time
 from contextvars import ContextVar
 from enum import Enum
 from functools import wraps
-from typing import Any, AsyncIterator, Mapping, Optional, Tuple
-from uuid import uuid4
+from typing import Optional, Tuple
 
-from cashews import LockedError
 from cashews._typing import AsyncCallable_T, Middleware
 from cashews.backends.interface import Backend
-from cashews.backends.memory import Memory
+from cashews.backends.transaction import LockTransactionBackend, TransactionBackend
 
 _transaction = ContextVar("transaction", default=None)
-_empty = object()
-_GLOBAL_LOCK_KEY = "cashews:serializable:lock"
 
 
 class TransactionMode(Enum):
     FAST = "fast"  # simple inmemory impl, allow to have lost updates,
     LOCKED = "locked"  # lock per key - not allow to work in parallel with the same key
     SERIALIZABLE = "serializable"  # global lock - not allow any parallel changes
-
-
-async def transaction(
-    mode: Optional[TransactionMode] = None, timeout: Optional[float] = None
-) -> "TransactionContextDecorator":
-    return TransactionContextDecorator(mode, timeout)
 
 
 class TransactionWrapperMixin:
@@ -54,9 +42,12 @@ class TransactionWrapperMixin:
 
 
 class TransactionContextDecorator:
+    __slots__ = ["_mode", "_timeout", "_inner"]
+
     def __init__(self, mode: Optional[TransactionMode] = None, timeout: Optional[float] = None):
         self._mode = mode
         self._timeout = timeout
+        self._inner = False
 
     @property
     def current_tx(self) -> Optional["Transaction"]:
@@ -64,7 +55,7 @@ class TransactionContextDecorator:
 
     async def __aenter__(self) -> "Transaction":
         if self.current_tx:
-            # warning that nested transaction not supported
+            self._inner = True
             return self.current_tx
         return self.start()
 
@@ -74,14 +65,14 @@ class TransactionContextDecorator:
         return tx
 
     async def __aexit__(self, exc_type, exc_value, exc_tb):
-        if not self.current_tx:
+        if not self.current_tx or self._inner:
+            self._inner = False
             return
         if not exc_tb:
-            await self.current_tx.commit()
+            await self.commit()
         else:
-            await self.current_tx.rollback()
+            await self.rollback()
         _transaction.set(None)
-        return False
 
     def __call__(self, func: AsyncCallable_T) -> AsyncCallable_T:
         @wraps(func)
@@ -101,6 +92,8 @@ class TransactionContextDecorator:
 
 
 class Transaction:
+    __slots__ = ["_mode", "_timeout", "_backends"]
+
     def __init__(self, mode: Optional[TransactionMode] = None, timeout: Optional[float] = None):
         self._mode = mode
         self._timeout = timeout
@@ -123,288 +116,3 @@ class Transaction:
     async def rollback(self):
         for tx_backend in self._backends.values():
             await tx_backend.rollback()
-
-
-class TransactionBackend(Backend):
-    def __init__(self, backend: Backend):
-        self._backend = backend
-        self._local_cache = Memory()
-        self._to_delete = set()
-        super().__init__()
-
-    def _key_is_delete(self, key: str) -> bool:
-        if key in self._to_delete:
-            return True
-        return False
-
-    async def commit(self):
-        if self._to_delete:
-            await self._backend.delete_many(*self._to_delete)
-
-        expire_group = {}
-        for key, (expire, value) in self._local_cache.store.items():
-            if expire:
-                expire = int(expire - time.time())
-            expire_group.setdefault(expire, {})[key] = value
-
-        for expire, kv in expire_group.items():
-            await self._backend.set_many(kv, expire=expire)
-        self._clear_local_storage()
-
-    async def rollback(self):
-        self._clear_local_storage()
-
-    def _clear_local_storage(self):
-        self._local_cache = Memory()
-        self._to_delete = set()
-
-    async def set(
-        self,
-        key: str,
-        value: Any,
-        expire: Optional[float] = None,
-        exist: Optional[bool] = None,
-    ) -> bool:
-        if exist is not None:
-            if not await self._backend.exists(key) is exist:
-                return False
-        if key in self._to_delete:
-            self._to_delete.remove(key)
-        return await self._local_cache.set(key, value, expire, exist)
-
-    async def set_many(self, pairs: Mapping[str, Any], expire: Optional[float] = None):
-        for key in pairs.keys():
-            if key in self._to_delete:
-                self._to_delete.remove(key)
-        return await self._local_cache.set_many(pairs, expire)
-
-    async def incr(self, key: str) -> int:
-        current = await self._backend.get(key, 0)
-        await self._local_cache.set(key, current + 1)
-        if key in self._to_delete:
-            self._to_delete.remove(key)
-        return current + 1
-
-    async def delete(self, key: str) -> bool:
-        await self._local_cache.delete(key)
-        self._to_delete.add(key)
-        return True
-
-    async def delete_many(self, *keys: str):
-        await self._local_cache.delete_many(*keys)
-        self._to_delete.update(keys)
-
-    async def delete_match(self, pattern: str):
-        await self._local_cache.delete_match(pattern)
-        async for key in self._backend.scan(pattern):
-            self._to_delete.add(key)
-
-    async def expire(self, key: str, timeout: float):
-        if self._key_is_delete(key):
-            return
-        value = await self._backend.get(key, default=_empty)
-        if value is _empty:
-            return await self._local_cache.expire(key, timeout)
-        await self._local_cache.set(key, value, expire=timeout)
-
-    # non transaction - proxy methods with custom logic
-    async def get(self, key: str, default: Optional[Any] = None) -> Any:
-        if self._key_is_delete(key):
-            return
-        value = await self._local_cache.get(key, default=_empty)
-        if value is not _empty:
-            return value
-        return await self._backend.get(key, default=default)
-
-    async def get_many(self, *keys: str, default: Optional[Any] = None) -> Tuple[Optional[Any], ...]:
-        missed_keys = set(keys)
-        values = {key: default for key in keys}
-
-        _keys = list(missed_keys)
-        for i, value in enumerate(await self._local_cache.get_many(*_keys, default=_empty)):
-            if value is not _empty:
-                key = _keys[i]
-                values[key] = value
-                missed_keys.remove(key)
-
-        _keys = list(missed_keys)
-        for i, value in enumerate(await self._backend.get_many(*_keys, default=default)):
-            key = _keys[i]
-            if not self._key_is_delete(key):
-                values[key] = value
-        return tuple(values[key] for key in keys)
-
-    async def get_match(self, pattern: str, batch_size: int = 100) -> AsyncIterator[Tuple[str, Any]]:
-        _local_state = set()
-        async for key, value in self._local_cache.get_match(pattern):
-            yield key, value
-            _local_state.add(key)
-        async for key, value in self._backend.get_match(pattern, batch_size=batch_size):
-            if self._key_is_delete(key):
-                continue
-            if key in _local_state:
-                continue
-            yield key, value
-
-    async def get_expire(self, key: str) -> int:
-        if self._key_is_delete(key):
-            return -2
-        local_expire = await self._local_cache.get_expire(key)
-        if local_expire >= 0:
-            return local_expire
-        return await self._backend.get_expire(key)
-
-    async def scan(self, pattern: str, batch_size: int = 100) -> AsyncIterator[str]:
-        _local_state = set()
-        async for key in self._local_cache.scan(pattern):
-            yield key
-            _local_state.add(key)
-        async for key in self._backend.scan(pattern, batch_size=batch_size):
-            if self._key_is_delete(key):
-                continue
-            if key in _local_state:
-                continue
-            yield key
-
-    async def exists(self, key: str) -> bool:
-        if await self._local_cache.exists(key):
-            return True
-        if self._key_is_delete(key):
-            return False
-        return await self._backend.exists(key)
-
-    # non transaction - proxy methods
-
-    @property
-    def is_init(self) -> bool:
-        return self._backend.is_init
-
-    async def init(self):
-        return await self._backend.init()
-
-    async def close(self):
-        return await self._backend.close()
-
-    async def set_raw(self, key: str, value: Any, **kwargs: Any):
-        return await self._backend.set_raw(key, value, **kwargs)
-
-    async def get_raw(self, key: str) -> Any:
-        return await self._backend.get_raw(key)
-
-    async def get_bits(self, key: str, *indexes: int, size: int = 1) -> Tuple[int, ...]:
-        return await self._backend.get_bits(key, *indexes, size=size)
-
-    async def incr_bits(self, key: str, *indexes: int, size: int = 1, by: int = 1) -> Tuple[int, ...]:
-        return await self._backend.incr_bits(key, *indexes, size=size, by=by)
-
-    async def slice_incr(self, key: str, start: int, end: int, maxvalue: int, expire: Optional[float] = None) -> int:
-        return await self._backend.slice_incr(key, start, end, maxvalue, expire)
-
-    async def get_size(self, key: str) -> int:
-        return await self._backend.get_size(key)
-
-    async def ping(self, message: Optional[bytes] = None) -> bytes:
-        return await self._backend.ping(message)
-
-    async def clear(self):
-        return await self._backend.clear()
-
-    async def set_lock(self, key: str, value: Any, expire: float) -> bool:
-        return await self._backend.set_lock(key, value, expire)
-
-    async def is_locked(
-        self,
-        key: str,
-        wait: Optional[float] = None,
-        step: float = 0.1,
-    ) -> bool:
-        return await self._backend.is_locked(key, wait, step)
-
-    async def unlock(self, key: str, value: Any) -> bool:
-        return await self._backend.unlock(key, value)
-
-
-class LockTransactionBackend(TransactionBackend):
-    def __init__(self, backend: Backend, serializable=False, timeout=10):
-        super().__init__(backend)
-        self._locks = set()
-        self._lock_id = uuid4().hex
-        self._serializable = serializable
-        self._timeout = timeout
-
-    def _get_lock_key(self, key: str) -> str:
-        if self._serializable:
-            return _GLOBAL_LOCK_KEY
-        return f":tx_lock:{key}"
-
-    async def _lock_updates(self, key: str):
-        lock_key = self._get_lock_key(key)
-        if lock_key in self._locks:
-            return
-        wait = self._timeout
-        step = 0.1
-        while wait > 0.0:
-            wait -= step
-            if await self._backend.set_lock(lock_key, self._lock_id, expire=self._timeout):
-                self._locks.add(lock_key)
-                return
-            await asyncio.sleep(step)
-        raise LockedError("probably deadlock or long running transactions")
-
-    async def _unlock_updates(self):
-        if self._locks:
-            await self._backend.delete_many(*self._locks)
-            self._locks = set()
-
-    async def commit(self):
-        try:
-            await super().commit()
-        finally:
-            await self._unlock_updates()
-
-    async def rollback(self):
-        try:
-            await super().rollback()
-        finally:
-            await self._unlock_updates()
-
-    async def set(
-        self,
-        key: str,
-        value: Any,
-        expire: Optional[float] = None,
-        exist: Optional[bool] = None,
-    ) -> bool:
-        await self._lock_updates(key)
-        return await super().set(key, value, expire=expire, exist=exist)
-
-    async def set_many(self, pairs: Mapping[str, Any], expire: Optional[float] = None):
-        for key in pairs.keys():
-            await self._lock_updates(key)
-        res = await super().set_many(pairs, expire=expire)
-        return res
-
-    async def incr(self, key: str) -> int:
-        await self._lock_updates(key)
-        return await super().incr(key)
-
-    async def delete(self, key: str) -> bool:
-        await self._lock_updates(key)
-        res = await super().delete(key)
-        return res
-
-    async def delete_many(self, *keys: str):
-        for key in keys:
-            await self._lock_updates(key)
-        res = await super().delete_many(*keys)
-        return res
-
-    async def delete_match(self, pattern: str):
-        await self._local_cache.delete_match(pattern)
-        async for key in self._backend.scan(pattern):
-            await self._lock_updates(key)
-            self._to_delete.add(key)
-
-    async def expire(self, key: str, timeout: float):
-        await self._lock_updates(key)
-        return await super().expire(key, timeout)
