+++ tests/test_middleware.py
@@ -1,27 +1,9 @@
-from unittest.mock import Mock
-
 import pytest
 
 from cashews import Cache
-from cashews.backends.memory import Memory
 from cashews.helpers import add_prefix, all_keys_lower, memory_limit
 
 pytestmark = pytest.mark.asyncio
-
-
-@pytest.fixture(name="target")
-def _target() -> Mock:
-    class MemCache(Memory):
-        pass
-
-    return Mock(wraps=MemCache())
-
-
-@pytest.fixture(name="cache")
-def __cache(target):
-    _cache = Cache()
-    _cache._add_backend(target)
-    return _cache
 
 
 async def test_all_keys_lower(cache: Cache, target):
